+++ Tests/ScaletorTests/ScaletorTests.swift
@@ -5,14 +5,14 @@ class ScaletorTests: XCTestCase {
     func test_Throws_WithDoubleFlat() {
         var error: Error?
         let note = Note(pitch: .a, accidental: .doubleFlat)
-        XCTAssertThrowsError(try Scaletor.generateScale(from: note, for: MajorMode.ionian)) { error = $0 }
+        XCTAssertThrowsError(try Scaletor.makeScale(from: note, for: MajorMode.ionian)) { error = $0 }
         XCTAssertEqual(error as? ScaleError, .theoreticalKey)
     }
 
     func test_Throws_WithDoubleSharp() {
         var error: Error?
         let note = Note(pitch: .a, accidental: .doubleSharp)
-        XCTAssertThrowsError(try Scaletor.generateScale(from: note, for: MajorMode.ionian)) { error = $0 }
+        XCTAssertThrowsError(try Scaletor.makeScale(from: note, for: MajorMode.ionian)) { error = $0 }
         XCTAssertEqual(error as? ScaleError, .theoreticalKey)
     }