+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 
 - Extract `ChordVoicing` description method to `ChordFormatter`
+- Rename `Scaletor.generateScale()` to `Scaletor.makeScale`
 
 ### Added